% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classification.r
\name{classification}
\alias{classification}
\title{Classify the given SpatialPointsDataFrame}
\usage{
classification(proj.df, useX = FALSE, lessType = FALSE, debug = FALSE)
}
\arguments{
\item{proj.df}{SpatialPointsDataFrame, the points to classify}

\item{useX}{boolean, use axe X for the classification}

\item{lessType}{boolean, merge similar types}

\item{debug}{boolean, to use the debug mode}
}
\value{
SpatialPointsDataFrame, the points classified
}
\description{
Classify the given SpatialPointsDataFrame
}
\author{
Valentin SASYAN, v. 2.2.0, 07/17/2015
}

