% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compileClassif.r
\name{compileClassif}
\alias{compileClassif}
\title{Copile the types given by the classification}
\usage{
compileClassif(proj.classif)
}
\arguments{
\item{proj.classif}{SpatialPointsDataFrame, the points classification}
}
\value{
proj.compile SpatialPointsDataFrame, the compiled classification
}
\description{
Copile the types given by the classification
}
\author{
Valentin SASYAN, v. 1.0.0, 07/28/2015
}

