% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExport.r
\name{convertJsonExport}
\alias{convertJsonExport}
\title{Convert the raw JSON export in a usable listPointXYZ}
\usage{
convertJsonExport(x, folder)
}
\arguments{
\item{x}{list, the raw JSON export}
}
\value{
list, the usable listPointXYZ
}
\description{
Convert the raw JSON export in a usable listPointXYZ
}
\author{
Valentin SASYAN, v. 1.2.0, 06/18/2015
}

