% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getASCinfo}
\alias{getASCinfo}
\title{Get information for all the .asc files of the folder given for parameter}
\usage{
getASCinfo(folder)
}
\arguments{
\item{folder}{string, folder to scan}
}
\value{
list, a list with all the information
}
\description{
Get information for all the .asc files of the folder given for parameter
}
\examples{
getASCinfo('data/generated/')
}
\author{
Valentin SASYAN, v. 1.0.1, 06/12/2015
}

