% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getContrastEnhancement}
\alias{getContrastEnhancement}
\title{Generate the 'ContrastEnhancement' XML element of a pipe XML element}
\usage{
getContrastEnhancement(color, liste)
}
\arguments{
\item{color}{string, name of the color}

\item{liste}{list, min and max for this color}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'ContrastEnhancement' XML element of a pipe XML element
}
\examples{
getContrastEnhancement('green', c(min=-0.84, max=0.42))
}
\author{
Valentin SASYAN, v. 1.0.0, 06/12/2015
}

