% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getEditType}
\alias{getEditType}
\title{Generate the 'edittype' XML element of a .shp 'maplayer' XML element}
\usage{
getEditType(editTypes)
}
\arguments{
\item{editTypes}{list, list of the name of the file's band}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'edittype' XML element of a .shp 'maplayer' XML element
}
\author{
Valentin SASYAN, v. 1.0.0, 06/12/2015
}

