% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getGoogle}
\alias{getGoogle}
\title{Create information for a Google Streets Layer}
\usage{
getGoogle(Google)
}
\arguments{
\item{Google}{boolean, must create the information or not?}
}
\value{
list, a list with all the information
}
\description{
Create information for a Google Streets Layer
}
\examples{
google <- getGoogle(TRUE)
google <- getGoogle(FALSE)
}
\author{
Valentin SASYAN, v. 1.0.0, 06/15/2015
}

