% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExport.r
\name{getIdType}
\alias{getIdType}
\title{Convert the String type of a building in a integer type}
\usage{
getIdType(typeString)
}
\arguments{
\item{typeString}{String, the String type of the building}
}
\value{
integer, the integer type of the building
}
\description{
Convert the String type of a building in a integer type
}
\author{
Valentin SASYAN, v. 1.0.0, 06/18/2015
}

