% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getLegendlayer}
\alias{getLegendlayer}
\title{Generate the 'noData' XML element of a 'legend' XML element}
\usage{
getLegendlayer(Name, ID)
}
\arguments{
\item{Name}{string, name of the layer}

\item{ID}{string, id of the layer}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'noData' XML element of a 'legend' XML element
}
\examples{
getLegendlayer('exportedData','exportedData20150612104242')
}
\author{
Valentin SASYAN, v. 1.0.0, 06/12/2015
}

