% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getMapcanvas}
\alias{getMapcanvas}
\title{Generate the 'mapcanvas' XML element of a 'qgis' XML element}
\usage{
getMapcanvas(xmin, ymin, xmax, ymax, EPSG)
}
\arguments{
\item{xmin}{double, xmin limit of the map canevas}

\item{ymin}{double, ymin limit of the map canevas}

\item{xmax}{double, xmax limit of the map canevas}

\item{ymax}{double, ymax limit of the map canevas}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'mapcanvas' XML element of a 'qgis' XML element
}
\examples{
getMapcanvas(-42,-42,42,42)
}
\author{
Valentin SASYAN, v. 1.1.0, 06/15/2015
}

