% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getMapLayerGoogle}
\alias{getMapLayerGoogle}
\title{Generate the 'maplayer' XML element of a Google Streets Layer}
\usage{
getMapLayerGoogle(id, layername, spatialrefsys)
}
\arguments{
\item{id}{string, id of the layer}

\item{layername}{string, layername of the layer}

\item{spatialrefsys}{XML element, specifications of the spatial reference systeme}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'maplayer' XML element of a Google Streets Layer
}
\author{
Valentin SASYAN, v. 1.0.0, 06/15/2015
}

