% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getMapLayerShp}
\alias{getMapLayerShp}
\title{Generate the 'maplayer' XML element of a .shp}
\usage{
getMapLayerShp(id, datasource, layername, spatialrefsys, editTypes)
}
\arguments{
\item{id}{string, id of the layer}

\item{datasource}{string, datasource of the layer}

\item{layername}{string, layername of the layer}

\item{spatialrefsys}{XML element, specifications of the spatial reference systeme}

\item{editTypes}{list, list of the name of the file's band}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'maplayer' XML element of a .shp
}
\author{
Valentin SASYAN, v. 1.0.0, 06/12/2015
}

