% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getNoData}
\alias{getNoData}
\title{Generate the 'noData' XML element of a .asc 'maplayer' XML element}
\usage{
getNoData(min, max)
}
\arguments{
\item{min}{int, first number of the data}

\item{max}{int, last number of the data}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'noData' XML element of a .asc 'maplayer' XML element
}
\examples{
getNoData(1,3)
}
\author{
Valentin SASYAN, v. 1.0.0, 06/12/2015
}

