% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getPipe}
\alias{getPipe}
\title{Generate the 'pipe' XML element of a .asc 'maplayer' XML element}
\usage{
getPipe(liste, opacity = 1)
}
\arguments{
\item{liste}{list, specification of the color used for each band}

\item{opacity}{double, opacity of the layer}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'pipe' XML element of a .asc 'maplayer' XML element
}
\examples{
getPipe(list(red=c(min=-1,max=1),green=c(min=-0.84, max=0.42),blue=c(min=-0.42,max=0.84))
}
\author{
Valentin SASYAN, v. 1.1.0, 06/15/2015
}

