% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolation.r
\name{getPointList}
\alias{getPointList}
\title{Generate a list of SpatialPoint}
\usage{
getPointList(extent, resolution)
}
\arguments{
\item{extent}{extent, area to cover}

\item{resolution}{double, resolution of the grid of interpolation (size of 1 px in real)}
}
\value{
list, a list of SpatialPoint
}
\description{
Generate a list of SpatialPoint
}
\author{
Valentin SASYAN, v. 1.1.0, 06/12/2015
}

