% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getRendererv2}
\alias{getRendererv2}
\title{Generate the 'renderer-v2' XML element of a .shp 'maplayer' XML element}
\usage{
getRendererv2()
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'renderer-v2' XML element of a .shp 'maplayer' XML element
}
\author{
Valentin SASYAN, v. 1.1.0, 06/22/2015
}

