% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getSHPinfo}
\alias{getSHPinfo}
\title{Get information for all the .shp files of the folder given for parameter}
\usage{
getSHPinfo(folder)
}
\arguments{
\item{folder}{string, folder to scan}
}
\value{
list, a list with all the information
}
\description{
Get information for all the .shp files of the folder given for parameter
}
\examples{
getSHPinfo('data/generated/')
}
\author{
Valentin SASYAN, v. 1.0.1, 06/12/2015
}

