% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getSpatialrefsys}
\alias{getSpatialrefsys}
\title{Generate the 'spatialrefsys' XML element of a 'maplayer' XML element}
\usage{
getSpatialrefsys(EPSG)
}
\arguments{
\item{EPSG}{list, EPSG description of the reference systeme used as destination}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'spatialrefsys' XML element of a 'maplayer' XML element
}
\author{
Valentin SASYAN, v. 1.1.0, 06/15/2015
}

