% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate_qgs.r
\name{getSymbol}
\alias{getSymbol}
\title{Generate the 'categroy' XML element of a .shp 'categories' XML element}
\usage{
getSymbol(i)
}
\arguments{
\item{i}{integer, indice of the symbol in the symbols list}
}
\value{
XMLNode, the XML element generated
}
\description{
Generate the 'categroy' XML element of a .shp 'categories' XML element
}
\author{
Valentin SASYAN, v. 1.1.0, 06/23/2015
}

