% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolation.r
\name{interpolation}
\alias{interpolation}
\title{Interpolate the date in the SpatialPointsDataFrame given for parameter}
\usage{
interpolation(proj.df, resolution = 100, EPSG, p = 0.5)
}
\arguments{
\item{proj.df}{SpatialPointsDataFrame, the data to interpolate}

\item{resolution}{double, resolution of the grid of interpolation (size of 1 px in real)}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}

\item{p}{real, power p for the interpolation}
}
\value{
SpatialPointsDataFrame, the data interpolated
}
\description{
Interpolate the date in the SpatialPointsDataFrame given for parameter
}
\author{
Valentin SASYAN, v. 1.4.0, 07/28/2015
}

