% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processData.r
\name{isWritable}
\alias{isWritable}
\title{Check if the script can write the specified file (erasing or not an eventual existing file)}
\usage{
isWritable(erase, file)
}
\arguments{
\item{erase}{boolean, the script can erase an eventual existing file}

\item{file}{string, path of the file to test}
}
\value{
boolean, true if the file can be written
}
\description{
Check if the script can write the specified file (erasing or not an eventual existing file)
}
\author{
Valentin SASYAN, v1.0.0, 06/12/2015
}

