% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExport.r
\docType{data}
\name{listTypeBuilding}
\alias{listTypeBuilding}
\title{List of the possible building type}
\format{\preformatted{ chr [1:16] "(W1) Wood, Light Frame (<= 5,000 sq. ft.)" ...
}}
\usage{
listTypeBuilding
}
\description{
List of the possible building type
}
\author{
LAGGISS, uOttawa
v. 1.1.1, 06/19/2015
}
\keyword{datasets}

