% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processData.r
\name{processData}
\alias{processData}
\title{Process (interpolation + classification) the data generated by the Android App and exported in the JSON file}
\usage{
processData(filter = "uOttawa", resolution = 0, export = TRUE,
  erase = TRUE, EPSG = EPSG_, interpolation = "idw", param = 0.5,
  classif = TRUE, qualify = FALSE, useX = FALSE, lessType = FALSE)
}
\arguments{
\item{filter}{RegExpr, filter for the name of the folders to use}

\item{resolution}{double, resolution of the grid of interpolation (size of 1 px in real), 0 for no interpolation}

\item{export}{bool, true if the function have to export the used data}

\item{erase}{bool, true if the function can erase existing used data}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}

\item{interpolation}{string, the name of the interplation to use}

\item{param}{variant, parameter for some interpolation}

\item{classif}{boolean, do or not the the classification}

\item{qualify}{boolean, show or not the classification statistics (needs classif=TRUE)}

\item{useX}{boolean, use axe X for the classification}

\item{lessType}{boolean, merge similar types}
}
\value{
nothing
}
\description{
Process (interpolation + classification) the data generated by the Android App and exported in the JSON file
}
\examples{
generate_qgs(filter='uOttawa.*',resolution=1,erase=FALSE)
}
\author{
Valentin SASYAN, v. 0.7.0, 07/22/2015
}

