% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qualifyClassif.r
\name{qualifyClassif}
\alias{qualifyClassif}
\title{Qualify the given classification}
\usage{
qualifyClassif(proj.classif)
}
\arguments{
\item{proj.classif}{SpatialPointsDataFrame, the points classification}
}
\value{
nothing
}
\description{
Qualify the given classification
}
\author{
Valentin SASYAN, v. 1.0.1, 07/22/2015
}

