% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readExport.r
\name{readExport}
\alias{readExport}
\title{Read the date in the json file}
\usage{
readExport(filter, EPSG)
}
\arguments{
\item{filter}{RegExpr, filter for the name of the folders to use}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}
}
\value{
SpatialPointsDataFrame, the data of the JSON in a SPDF ready for the interpolation
}
\description{
Read the date in the json file
}
\author{
Valentin SASYAN, v. 1.1.0, 06/16/2015
}

