% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/processData.r
\name{sizeData}
\alias{sizeData}
\title{Calculate the size of the interpolation mesh for the given data and the given interpolation}
\usage{
sizeData(file = "generated", resolution = 1000, EPSG = EPSG_)
}
\arguments{
\item{file}{string, name of the JSON file in the data folder (without the .json extension)}

\item{resolution}{double, resolution of the grid of interpolation (size of 1 px in real)}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}
}
\value{
list, the X size and the Y size
}
\description{
Calculate the size of the interpolation mesh for the given data and the given interpolation
}
\examples{
sizeData(file'generated',resolution=1)
}
\author{
Valentin SASYAN, v. 0.2.0, 06/15/2015
}

