% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolation.r
\name{interpolation}
\alias{interpolation}
\title{Interpolate the date in the SpatialPointsDataFrame given for parameter}
\usage{
interpolation(proj.df, resolution = 100, EPSG, interpolation = "idw",
  param = 0.5)
}
\arguments{
\item{proj.df}{SpatialPointsDataFrame, the data to interpolate}

\item{resolution}{double, resolution of the grid of interpolation (size of 1 px in real)}

\item{EPSG}{list, EPSG description of the reference systeme used as destination}

\item{interpolation}{string, the name of the interplation to use}

\item{param}{variant, parameter for some interpolation}
}
\value{
SpatialPointsDataFrame, the data interpolated
}
\description{
Interpolate the date in the SpatialPointsDataFrame given for parameter
}
\author{
Valentin SASYAN, v. 1.3.1, 06/17/2015
}

